+++ src/modules/getdata.js
@@ -1,0 +1,9 @@
+
+const getData = async () => {
+    const response = await fetch (url)
+    const data = await response.json()
+    // console.log(data)
+    return data.result
+}
+
+export { getData }
